Rename ATLControl files to ComServer to harmonize naming.
