+++ ComServer/ComServer.vcxproj
@@ -124,14 +124,14 @@
     </Link>
   </ItemDefinitionGroup>
   <ItemGroup>
-    <ClInclude Include="ATLControl.h" />
+    <ClInclude Include="ComServer.h" />
     <ClInclude Include="framework.h" />
     <ClInclude Include="pch.h" />
     <ClInclude Include="Resource.h" />
     <ClInclude Include="targetver.h" />
   </ItemGroup>
   <ItemGroup>
-    <ClCompile Include="ATLControl.cpp" />
+    <ClCompile Include="ComServer.cpp" />
     <ClCompile Include="ComServerModule.cpp" />
     <ClCompile Include="pch.cpp">
       <PrecompiledHeader Condition="'$(Configuration)|$(Platform)'=='Debug|x64'">Create</PrecompiledHeader>
@@ -148,9 +148,6 @@
   </ItemGroup>
   <ItemGroup>
     <Midl Include="ComServerModule.idl" />
-  </ItemGroup>
-  <ItemGroup>
-    <Image Include="ATLControl.bmp" />
   </ItemGroup>
   <Import Project="$(VCTargetsPath)\Microsoft.Cpp.targets" />
   <ImportGroup Label="ExtensionTargets">
